QP Update & Replace Player (#26)
* Init

* replace-player command

* added checkAdmin() to some commands

* QP fix

* small chnages

+++ serverInstance.py
@@ -56,9 +56,6 @@ class serverInstance:
 				discordUser = None
 			player = Player(player_details[0], player_details[1], player_details[2], player_details[3], player_details[4], player_details[5], player_details[6], player_details[7], player_details[8],
                    player_details[9], player_details[10], player_details[11], self.cursor, self.con, discordUser)
-			# Add signup
-			player.addSignUpCount()
-			player.update()
 			# Add player to list
 			playerObjList.append(player)
    
@@ -528,33 +525,38 @@ After a win, post a screenshot of the victory and type !win (only one player on 
 	# Method to End a Current Match if not started or void
 	async def endMatch(self, message_obj, matchID):
      
-		# Check if matchID is in current matches
-		for match in self.currentMatches:
-			match_id = match.get_matchID()
-			try:
-				if match_id == int(matchID):
-					# Delete Match
-					match.delete()
-					# Pop match off list
-					self.currentMatches.remove(match)
-					await message_obj.channel.send(f"🗑️ Match ({match_id}) Removed")
-			except:
-				pass
+		# Check if admin
+		admin_check = await self.checkAdmin(message_obj.author.id)
+		if admin_check:
+			# Check if matchID is in current matches
+			for match in self.currentMatches:
+				match_id = match.get_matchID()
+				try:
+					if match_id == int(matchID):
+						# Delete Match
+						match.delete()
+						# Pop match off list
+						self.currentMatches.remove(match)
+						await message_obj.channel.send(f"🗑️ Match ({match_id}) Removed")
+				except:
+					pass
 
 	# Method to punish a player -> reducing LP and QP
 	async def punishPlayer(self, message_obj, discordID):
 		
-		
-		discordID = discordID.replace("<@", "")
-		discordID = discordID.replace(">", "")
-		res = self.cursor.execute(f"SELECT leaderboardPoints, signupCount, discordID FROM Player WHERE discordID = {discordID}")
-		result = res.fetchone()
-		new_mmr = result[0] - 50
-		add_signupCount = result[1] + 3
-		user = await self.client.fetch_user(discordID)
-		self.cursor.execute(f"UPDATE Player SET leaderboardPoints = {new_mmr}, signupCount = {add_signupCount} WHERE discordID = {discordID}")
-		self.con.commit()
-		await message_obj.channel.send(f"🔨 {user.mention} has been given a pentaly of -50**LP** and added to **Low Priority Queue**")
+		# Check if admin
+		admin_check = await self.checkAdmin(message_obj.author.id)
+		if admin_check:
+			discordID = discordID.replace("<@", "")
+			discordID = discordID.replace(">", "")
+			res = self.cursor.execute(f"SELECT leaderboardPoints, signupCount, discordID FROM Player WHERE discordID = {discordID}")
+			result = res.fetchone()
+			new_mmr = result[0] - 50
+			add_signupCount = result[1] + 3
+			user = await self.client.fetch_user(discordID)
+			self.cursor.execute(f"UPDATE Player SET leaderboardPoints = {new_mmr}, signupCount = {add_signupCount} WHERE discordID = {discordID}")
+			self.con.commit()
+			await message_obj.channel.send(f"🔨 {user.mention} has been given a pentaly of -50**LP** and added to **Low Priority Queue**")
 		
 	# Method to swap two players on the same team
 	async def swapPlayers(self, message_obj, discordIDOtherPlayer):
@@ -577,6 +579,37 @@ After a win, post a screenshot of the victory and type !win (only one player on 
 						pass
 		except asyncio.TimeoutError:
 			await message_obj.channel.send(f"Swap timed out {message_obj.author.mention}")
+   
+   
+	async def replacePlayer(self, msg_obj, discordIDOrigin, discordIDReplacement):
+     
+		# Check called by (isAdmin)
+		user_id = msg_obj.author.id
+		admin_check = await self.checkAdmin(user_id)
+		if admin_check:
+			# Parse discord ID's from mentions in message
+			discordIDOrigin = discordIDOrigin.replace("<@", "")
+			discordIDOrigin = discordIDOrigin.replace(">", "")
+			discordIDReplacement = discordIDReplacement.replace("<@", "")
+			discordIDReplacement = discordIDReplacement.replace(">", "")
+			# Check for players in every match -> if found, replace player
+			for match in self.currentMatches:
+				try:
+					await match.swapPlayer(discordIDOrigin, discordIDReplacement, msg_obj)
+				except:
+					await msg_obj.channel.send(f"Replacement Error")
+		else:
+			pass
+			
+		
+	# Check if discordID is Admin
+	async def checkAdmin(self, discordID):
+		res = self.cursor.execute(f"SELECT isAdmin FROM Player WHERE discordID = {discordID}")
+		result = res.fetchone()
+		if(result[0] == 1):
+			return True
+		else:
+			return False
